+++ packages/react-reconciler/src/ReactFiberBeginWork.js
@@ -31,7 +31,7 @@ import {
   ContextConsumer,
   Profiler,
   PlaceholderComponent,
-} from 'shared/ReactTypeOfWork';
+} from 'shared/ReactWorkTags';
 import {
   NoEffect,
   PerformedWork,
@@ -40,7 +40,7 @@ import {
   DidCapture,
   Update,
   Ref,
-} from 'shared/ReactTypeOfSideEffect';
+} from 'shared/ReactSideEffectTags';
 import ReactSharedInternals from 'shared/ReactSharedInternals';
 import {
   enableGetDerivedStateFromCatch,
