+++ packages/react-reconciler/src/ReactFiberClassComponent.js
@@ -11,7 +11,7 @@ import type {Fiber} from './ReactFiber';
 import type {ExpirationTime} from './ReactFiberExpirationTime';
 
 import React from 'react';
-import {Update, Snapshot} from 'shared/ReactTypeOfSideEffect';
+import {Update, Snapshot} from 'shared/ReactSideEffectTags';
 import {
   debugRenderPhaseSideEffects,
   debugRenderPhaseSideEffectsForStrictMode,
