+++ packages/react-reconciler/src/ReactFiberHydrationContext.js
@@ -16,8 +16,8 @@ import type {
   HostContext,
 } from './ReactFiberHostConfig';
 
-import {HostComponent, HostText, HostRoot} from 'shared/ReactTypeOfWork';
-import {Deletion, Placement} from 'shared/ReactTypeOfSideEffect';
+import {HostComponent, HostText, HostRoot} from 'shared/ReactWorkTags';
+import {Deletion, Placement} from 'shared/ReactSideEffectTags';
 import invariant from 'shared/invariant';
 
 import {createFiberFromHostInstanceForDeletion} from './ReactFiber';
