+++ packages/shared/ReactSideEffectTags.js
@@ -7,7 +7,7 @@
  * @flow
  */
 
-export type TypeOfSideEffect = number;
+export type SideEffectTag = number;
 
 // Don't change these two values. They're used by React Dev Tools.
 export const NoEffect = /*              */ 0b00000000000;
